Add files via upload
To create a basic application to check if a string is accepted by a DFA or not. The entire application is a terminal line application with a graph which visualises the DFA form in terms of a directed graph. 
The basic idea is to make the process of creating and visualizing the DFA easier and the tool helps to examine if the string is accepted or not by the given deterministic automata. The tool can visualize huge DFA which is very tedious pen paper process.